requirements.txt
requests==2.32.3
feedparser==6.0.11
APScheduler==3.10.4
anthropic==0.39.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
